+++ manage_breast_screening/mammograms/presenters/appointment_presenters.py
@@ -2,25 +2,9 @@ from functools import cached_property
 
 from django.urls import reverse
 
-from ..core.utils.date_formatting import format_date, format_relative_date, format_time
-from ..participants.models import AppointmentStatus, SupportReasons
-from ..participants.presenters import ParticipantPresenter, status_colour
-
-
-def present_secondary_nav(pk):
-    """
-    Build a secondary nav for reviewing the information of screened/partially screened appointments.
-    """
-    return [
-        {
-            "id": "all",
-            "text": "Appointment details",
-            "href": reverse("mammograms:show_appointment", kwargs={"pk": pk}),
-            "current": True,
-        },
-        {"id": "medical_information", "text": "Medical information", "href": "#"},
-        {"id": "images", "text": "Images", "href": "#"},
-    ]
+from ...core.utils.date_formatting import format_date, format_relative_date, format_time
+from ...participants.models import AppointmentStatus, SupportReasons
+from ...participants.presenters import ParticipantPresenter, status_colour
 
 
 class AppointmentPresenter:
@@ -91,76 +75,6 @@ class AppointmentPresenter:
             "key": current_status.state,
             "is_confirmed": current_status.state == AppointmentStatus.CONFIRMED,
             "is_screened": current_status.state == AppointmentStatus.SCREENED,
-        }
-
-
-class LastKnownMammogramPresenter:
-    def __init__(self, last_known_mammograms, participant_pk, current_url):
-        self._last_known_mammograms = last_known_mammograms
-        self.participant_pk = participant_pk
-        self.current_url = current_url
-
-    @cached_property
-    def last_known_mammograms(self):
-        result = []
-        for mammogram in self._last_known_mammograms:
-            result.append(self._present_mammogram(mammogram))
-
-        return result
-
-    def _present_mammogram(self, mammogram):
-        location = (
-            mammogram.provider.name
-            if mammogram.provider
-            else mammogram.location_details
-        )
-        if mammogram.provider:
-            location = mammogram.provider.name
-        elif (
-            mammogram.location_details
-            and mammogram.location_type == mammogram.LocationType.ELSEWHERE_UK
-        ):
-            location = f"In the UK: {mammogram.location_details}"
-        elif (
-            mammogram.location_details
-            and mammogram.location_type == mammogram.LocationType.OUTSIDE_UK
-        ):
-            location = f"Outside the UK: {mammogram.location_details}"
-        elif mammogram.location_type == mammogram.LocationType.PREFER_NOT_TO_SAY:
-            location = "Location: prefer not to say"
-        else:
-            location = "Location unknown"
-
-        if mammogram.exact_date:
-            absolute = format_date(mammogram.exact_date)
-            relative = format_relative_date(mammogram.exact_date)
-            date = {"absolute": absolute, "relative": relative, "is_exact": True}
-        elif mammogram.approx_date:
-            date = {"value": f"Approximate date: {mammogram.approx_date}"}
-        else:
-            date = {"value": "Date unknown"}
-
-        return {
-            "date_added": format_relative_date(mammogram.created_at),
-            "location": location,
-            "date": date,
-            "different_name": mammogram.different_name,
-            "additional_information": mammogram.additional_information,
-        }
-
-    @cached_property
-    def add_link(self):
-        href = (
-            reverse(
-                "participants:add_previous_mammogram",
-                kwargs={"pk": self.participant_pk},
-            )
-            + f"?return_url={self.current_url}"
-        )
-        return {
-            "href": href,
-            "text": "Add another" if self.last_known_mammograms else "Add",
-            "visually_hidden_text": "mammogram",
         }
